requirements.txt
Flask==1.1.2
pytest
pylint
google-cloud-storage==1.24.1